Split phone number step to reduce bundle for other registration pages
